Modify project for use of compatible build class
Using `AbstractBuild` class in the Actions is not working for
the Workflow (pipeline) builds. Considering that we do not need
any of the functionality of a more specialized class, we replaced
the use of this class with `Run`.

This plug-in now properly records the results in a step. It also
shows a link to a table with the collected metrics in Jenkins.